🏗️ Tore `tuple` measurement into pieces.
