+++ test/unit/tuple/unit.test.ts
@@ -1,0 +1,8 @@
+import {Measurement, unit} from "!src/tuple";
+
+describe('unit', () => {
+  it("should return the unit of the measurement.", () => {
+    const measurement: Measurement<"px", 12.5> = [12.5, "px"] as const;
+    expect(unit(measurement)).toBe("px");
+  });
+});
